Moved test methods
